+++ test/host-shim/atn.h
@@ -195,6 +195,30 @@ typedef struct stateAnyFalseFb
 	plcbit value;
 } stateAnyFalseFb_typ;
 
+typedef struct valueRefFb
+{
+	/* VAR_INPUT (analog) */
+	plcstring state[81];
+	unsigned long sData;
+	plcstring owner[81];
+	unsigned long pStatus;
+	unsigned long sStatus;
+	/* VAR_OUTPUT (analog) */
+	unsigned long data;
+	/* VAR (analog) */
+	unsigned long* cache;
+	plcstring registeredTopic[81];
+	/* VAR_INPUT (digital) */
+	plcbit update;
+	/* VAR_OUTPUT (digital) */
+	plcbit bound;
+	plcbit valid;
+	plcbit sizeMismatch;
+	plcbit returnBound;
+	/* VAR (digital) */
+	plcbit registered;
+} valueRefFb_typ;
+
 typedef struct AtnPLCOpen
 {
 	/* VAR_INPUT (analog) */
@@ -259,8 +283,14 @@ _BUR_PUBLIC unsigned long registerStateBoolAdr(plcstring* state, plcstring* modu
 _BUR_PUBLIC unsigned long registerStateParameters(plcstring* state, plcstring* moduleName, unsigned long* pParameters, unsigned long sParameters);
 _BUR_PUBLIC unsigned long registerStateExt1(plcstring* state, plcstring* moduleName, plcstring* pModuleStatus, unsigned long* pParameters, unsigned long sParameters, plcbit* pModuleByPass, plcbit* pActive);
 _BUR_PUBLIC unsigned long registerToResource(plcstring* resource, plcstring* moduleName, unsigned long* pResourceUserId, plcbit* pResourceActive);
-_BUR_PUBLIC unsigned long unregister(plcstring* name, plcstring* owner);
-_BUR_PUBLIC unsigned long unregisterAll(plcstring* owner);
+_BUR_PUBLIC unsigned long registerValue(plcstring* state, plcstring* owner, unsigned long* pData, unsigned long sData, plcbit* valid, unsigned long sReturn, unsigned long returnTopic);
+_BUR_PUBLIC void valueRefFb(struct valueRefFb* inst);
+_BUR_PUBLIC unsigned long unregister(plcstring* name);
+_BUR_PUBLIC unsigned long unregisterAll(void);
+/* Host-shim only, not part of the AS library interface: overrides the task name
+   reported by atnCurrentTaskName() for the calling thread, so a single-threaded
+   test can simulate multiple tasks. Pass 0 or "" to revert to the thread-id name. */
+_BUR_PUBLIC void atnSetCurrentTaskName(const char* name);
 _BUR_PUBLIC unsigned long subscribeCommandBool(plcstring* commandName, plcstring* moduleName, plcbit* command);
 _BUR_PUBLIC unsigned long subscribePLCOpen(plcstring* commandName, plcstring* moduleName, plcbit* command, struct AtnPlcOpenStatus* status);
 _BUR_PUBLIC unsigned long subscribePLCOpenWithParameters(plcstring* commandName, plcstring* moduleName, unsigned long* pParameters, unsigned long sParameters, plcbit* command, struct AtnPlcOpenStatus* status);
@@ -287,8 +317,10 @@ _BUR_PUBLIC signed long atnSetDiagnosticLogger(plcstring* loggerName);
 /* Constants */
 #ifdef _REPLACE_CONST
  #define ATN_ACTION_NAME_LEN 20U
+ #define ATN_RETURN_TOPIC_SUFFIX "~return"
 #else
  _GLOBAL_CONST unsigned char ATN_ACTION_NAME_LEN;
+ _GLOBAL_CONST plcstring ATN_RETURN_TOPIC_SUFFIX[8];
 #endif